requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiofiles==23.2.1
    # via gcloud-aio-storage
aiohttp==3.9.1
    # via
    #   gcloud-aio-auth
    #   gundi-client
aioredis==2.0.1
    # via -r requirements.in
aiosignal==1.3.1
    # via aiohttp
anyio==4.2.0
    # via
    #   httpcore
    #   starlette
async-timeout==4.0.3
    # via
    #   aiohttp
    #   aioredis
attrs==23.2.0
    # via aiohttp
backoff==2.2.0
    # via
    #   -r requirements.in
    #   gcloud-aio-auth
    #   opentelemetry-exporter-otlp-proto-http
backports-zoneinfo==0.2.1
    # via tzlocal
cachetools==5.3.2
    # via google-auth
certifi==2023.11.17
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
chardet==5.2.0
    # via gcloud-aio-auth
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
cryptography==41.0.7
    # via gcloud-aio-auth
dateparser==1.2.0
    # via smartconnect-client
defusedxml==0.7.1
    # via untangle
deprecated==1.2.14
    # via opentelemetry-api
environs==9.5.0
    # via
    #   -r requirements.in
    #   gundi-client
    #   gundi-client-v2
    #   smartconnect-client
exceptiongroup==1.2.0
    # via anyio
fastapi==0.108.0
    # via -r requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
gcloud-aio-auth==5.1.2
    # via
    #   gcloud-aio-pubsub
    #   gcloud-aio-storage
gcloud-aio-pubsub==6.0.0
    # via -r requirements.in
gcloud-aio-storage==9.2.0
    # via -r requirements.in
google-api-core[grpc]==2.15.0
    # via
    #   google-api-core
    #   google-cloud-trace
google-auth==2.26.1
    # via google-api-core
google-cloud-trace==1.12.0
    # via opentelemetry-exporter-gcp-trace
googleapis-common-protos==1.62.0
    # via
    #   google-api-core
    #   grpcio-status
    #   opentelemetry-exporter-otlp-proto-http
grpcio==1.60.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.60.0
    # via google-api-core
gundi-client==0.2.2
    # via -r requirements.in
gundi-client-v2==2.3.3
    # via -r requirements.in
gundi-core==1.5.9
    # via
    #   -r requirements.in
    #   gundi-client-v2
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
hiredis==2.3.2
    # via -r requirements.in
httpcore==0.17.3
    # via httpx
httpx==0.24.1
    # via
    #   -r requirements.in
    #   gundi-client-v2
    #   respx
    #   smartconnect-client
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
marshmallow==3.20.1
    # via environs
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
numpy==1.21.6
    # via
    #   shapely
    #   smartconnect-client
    #   timezonefinder
opentelemetry-api==1.14.0
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-gcp-trace
    #   opentelemetry-exporter-otlp-proto-http
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-propagator-gcp
    #   opentelemetry-sdk
opentelemetry-exporter-gcp-trace==1.3.0
    # via -r requirements.in
opentelemetry-exporter-otlp-proto-http==1.15.0
    # via -r requirements.in
opentelemetry-instrumentation==0.35b0
    # via
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-instrumentation-requests
opentelemetry-instrumentation-aiohttp-client==0.35b0
    # via -r requirements.in
opentelemetry-instrumentation-httpx==0.35b0
    # via -r requirements.in
opentelemetry-instrumentation-requests==0.35b0
    # via -r requirements.in
opentelemetry-propagator-gcp==1.6.0
    # via -r requirements.in
opentelemetry-proto==1.15.0
    # via opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk==1.14.0
    # via
    #   -r requirements.in
    #   opentelemetry-exporter-gcp-trace
    #   opentelemetry-exporter-otlp-proto-http
opentelemetry-semantic-conventions==0.35b0
    # via
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-httpx
    #   opentelemetry-instrumentation-requests
    #   opentelemetry-sdk
opentelemetry-util-http==0.35b0
    # via
    #   opentelemetry-instrumentation-aiohttp-client
    #   opentelemetry-instrumentation-requests
packaging==23.2
    # via marshmallow
prometheus-client==0.19.0
    # via gcloud-aio-pubsub
proto-plus==1.23.0
    # via google-cloud-trace
protobuf==4.25.1
    # via
    #   google-api-core
    #   google-cloud-trace
    #   googleapis-common-protos
    #   grpcio-status
    #   opentelemetry-proto
    #   proto-plus
pyasn1==0.5.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via
    #   gcloud-aio-storage
    #   google-auth
pycparser==2.21
    # via cffi
pydantic==1.10.13
    # via
    #   -r requirements.in
    #   fastapi
    #   gundi-client
    #   gundi-client-v2
    #   gundi-core
    #   smartconnect-client
pyjwt==2.8.0
    # via gcloud-aio-auth
python-dateutil==2.8.2
    # via dateparser
python-dotenv==1.0.0
    # via environs
python-json-logger==2.0.7
    # via smartconnect-client
pytz==2021.3
    # via
    #   dateparser
    #   smartconnect-client
redis==3.5.3
    # via
    #   smartconnect-client
    #   walrus
regex==2023.12.25
    # via dateparser
requests==2.31.0
    # via
    #   google-api-core
    #   opentelemetry-exporter-otlp-proto-http
    #   smartconnect-client
respx==0.20.2
    # via gundi-client-v2
rsa==4.9
    # via
    #   gcloud-aio-storage
    #   google-auth
shapely==2.0.2
    # via smartconnect-client
six==1.16.0
    # via python-dateutil
smartconnect-client @ https://github.com/PADAS/smartconnect-client/releases/download/v1.5.3/smartconnect_client-1.5.3-py3-none-any.whl
    # via -r requirements.in
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.32.0.post1
    # via fastapi
timezonefinder==5.2.0
    # via smartconnect-client
typing-extensions==4.9.0
    # via
    #   aioredis
    #   anyio
    #   fastapi
    #   opentelemetry-sdk
    #   pydantic
    #   starlette
    #   uvicorn
tzlocal==5.2
    # via dateparser
untangle==1.2.1
    # via smartconnect-client
urllib3==2.1.0
    # via requests
uvicorn==0.25.0
    # via -r requirements.in
walrus==0.9.2
    # via -r requirements.in
wrapt==1.16.0
    # via
    #   deprecated
    #   opentelemetry-instrumentation
    #   opentelemetry-instrumentation-aiohttp-client
yarl==1.9.4
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
